Add generic and raw service (#28)
* Two more services added

* Adding description

* Add more field descriptions

The raw service requires the serial number for your appliance and willl allow you to construct a payload without validation in the integration. However, each service must fulfill some preconditions. These are enforced by the API at Miele and error messages will be logged if the API did not accept the service call.

The generic service has some more validation built in but is considered as Work In Progress.

All services can be tested on developer tools page.

Closes #25